requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
reportlab==4.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
PyMuPDF==1.23.8
Pillow==10.2.0
pytest==8.0.0
flake8==7.0.0 